requirements.txt
python-xlib
inotify_simple
psutil
git+https://github.com/luizoti/evdev.git